third_party/blink/web_tests/external/wpt/storage-access-api/storage-access-beyond-cookies.BlobURLSharedWorker.sub.https.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify a blob URL created via URL.createObjectURL from the third-party context shouldn't be useable as the shared worker script when it's passed to saa_handle.SharedWorker.
  assert_equals: expected "Blob URL SharedWorker tests completed successfully." but got "First-party SharedWorker created from third-party blob URL unexpectedly."
Harness: the test ran to completion.
